Move WB methods to the view
