+++ workbaskets/models.py
@@ -14,14 +14,9 @@ from django.contrib.auth import get_user_model
 from django.core.exceptions import ImproperlyConfigured
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.db.models import Case
-from django.db.models import DateField
-from django.db.models import F
 from django.db.models import Max
 from django.db.models import QuerySet
 from django.db.models import Subquery
-from django.db.models import Value
-from django.db.models import When
 from django_fsm import FSMField
 from django_fsm import transition
 
@@ -751,83 +746,6 @@ class WorkBasket(TimestampedMixin):
     def assigned_reviewers(self):
         user_ids = self.reviewer_assignments.values_list("user_id", flat=True)
         return User.objects.filter(id__in=user_ids)
-
-    def get_measures_to_end_date(self) -> QuerySet:
-        """
-        Returns a queryset of measures on end-dated commodities in the
-        workbasket along with those commodities' end-dates.
-
-        It filters out measures which have already ended.
-        """
-
-        from commodities.models.orm import GoodsNomenclature
-
-        end_dated_commodities = GoodsNomenclature.objects.current().filter(
-            transaction__workbasket=self,
-            valid_between__upper_inf=False,
-        )
-        commodity_dict = {
-            commodity.sid: commodity.valid_between
-            for commodity in end_dated_commodities
-        }
-        measures_on_commodities = Measure.objects.current().filter(
-            goods_nomenclature__sid__in=commodity_dict.keys(),
-        )
-        conditions = [
-            When(
-                goods_nomenclature__sid=commodity_sid,
-                then=Value(commodity_valid_between),
-            )
-            for commodity_sid, commodity_valid_between in commodity_dict.items()
-        ]
-        measures = measures_on_commodities.annotate(
-            commodity_valid_between=Case(
-                *conditions,
-                output_field=DateField(),
-            ),
-        )
-
-        return measures.with_effective_valid_between().exclude(
-            db_effective_valid_between__not_gt=F("commodity_valid_between"),
-        )
-
-    def get_footnote_associations_to_end_date(self):
-        """Queries all end-dated commodites in a workbasket and finds related
-        footnote associations so they can be end-dated too."""
-        from commodities.models.orm import FootnoteAssociationGoodsNomenclature
-        from commodities.models.orm import GoodsNomenclature
-
-        end_dated_commodities = GoodsNomenclature.objects.current().filter(
-            transaction__workbasket=self,
-            valid_between__upper_inf=False,
-        )
-        commodity_dict = {
-            commodity.sid: commodity.valid_between
-            for commodity in end_dated_commodities
-        }
-
-        footnote_associations = (
-            FootnoteAssociationGoodsNomenclature.objects.current().filter(
-                goods_nomenclature__sid__in=commodity_dict.keys(),
-            )
-        )
-        conditions = [
-            When(
-                goods_nomenclature__sid=commodity_sid,
-                then=Value(commodity_valid_between),
-            )
-            for commodity_sid, commodity_valid_between in commodity_dict.items()
-        ]
-        footnote_associations = footnote_associations.annotate(
-            commodity_valid_between=Case(
-                *conditions,
-                output_field=DateField(),
-            ),
-        )
-
-        return footnote_associations.exclude(
-            valid_between__not_gt=F("commodity_valid_between"),
-        )
 
     class Meta:
         verbose_name = "workbasket"
